Setup/requirements.txt
Jinja2==2.8
requests==2.10.0
PyYAML==3.11
